[UPDT] - Update Hyprland configuration to v0.53.0 format
The update to Hyprland v0.53.0 breaking made changes to the window rule syntax
and values. To update the configuration to the new format the window
rules will be adjusted in line with the example window rules from
Hyprland. See https://github.com/hyprwm/Hyprland/releases/tag/v0.53.0
for more details.

Signed-off-by: Eichenour, Riley <R.D.Eichenour@gmail.com>